requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
duckduckgo_search==7.3.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
gradio==5.16.0
gradio_client==1.7.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.61.1
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
networkx==3.4.2
numpy==2.2.2
ollama==0.4.7
openai==1.62.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
primp==0.12.1
propcache==0.2.1
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smolagents==1.8.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==14.2
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
